adjusted some implicit casts

+++ CSP/Crypto/Base64.cpp
@@ -95,10 +95,10 @@ CBase64::~CBase64()
 std::string &CBase64::Encode(ByteArray &data, std::string &encodedData) {
 
 	init_func
-	DWORD dwStrSize = 0;
-	CryptBinaryToString(data.data(), (DWORD)data.size(), CRYPT_STRING_BASE64, NULL, &dwStrSize);
+	size_t dwStrSize = 0;
+	CryptBinaryToString(data.data(), (size_t)data.size(), CRYPT_STRING_BASE64, NULL, &dwStrSize);
 	encodedData.resize(dwStrSize);
-	CryptBinaryToString(data.data(), (DWORD)data.size(), CRYPT_STRING_BASE64, &encodedData.front(), &dwStrSize);
+	CryptBinaryToString(data.data(), (size_t)data.size(), CRYPT_STRING_BASE64, &encodedData.front(), &dwStrSize);
 
 	return encodedData;
 	exit_func
@@ -107,7 +107,7 @@ std::string &CBase64::Encode(ByteArray &data, std::string &encodedData) {
 ByteDynArray &CBase64::Decode(const char *encodedData,ByteDynArray &data) {
 	init_func
 
-	DWORD dwDataSize = 0;
+	size_t dwDataSize = 0;
 	CryptStringToBinary(encodedData, 0, CRYPT_STRING_BASE64, NULL, &dwDataSize, NULL, NULL);
 	data.resize(dwDataSize);
 	CryptStringToBinary(encodedData, 0, CRYPT_STRING_BASE64, data.data(), &dwDataSize, NULL, NULL);
